backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
google-genai
requests
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
sqlalchemy
aiosqlite
paho-mqtt
